Separte out tests for different endpoints to different files
